chore($git): first commit
This first commit already contains a working version of the
gem, however most of the configuration in the hosting
project is currently done by hand.

The engine declares the `spectacular` routes namespace
where the index render the test runner and the `fixtures`
subpath gives access to the fixtures defined in the
`fixtures_path` setting.

There's a really quick and dirty handling of fixtures types
by format comparison.